refactor(filtering): fix in rendering, redundant code removed
